// frontend/src/api/model/workoutPlanRead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Gym Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanExerciseRead } from './planExerciseRead';

export interface WorkoutPlanRead {
  id: number;
  name: string;
  is_active: boolean;
  created_at: string;
  updated_at: string;
  exercises?: PlanExerciseRead[];
}